gather game loop into its corresponding struct
